+++ Core/Src/audio.c
@@ -44,7 +44,7 @@ void audio_ProcessDMABuffer(uint16_t dmaBufferStart) {
 	audioBufferUpdated = 1;
 }
 
-int16_t* audio_GetBuffer() {
+uint8_t* audio_GetBuffer() {
 	audioBufferUpdated = 0;
 	return &audioBuffer[0];
 }
@@ -52,6 +52,4 @@ int16_t* audio_GetBuffer() {
 uint8_t* audio_BufferUpdated() {
 	return &audioBufferUpdated;
 }
-
-
